requirements.txt
langchain~=0.0.198
openai~=0.27.8
docarray
tiktoken
python-dotenv~=1.0.0
streamlit~=1.24.0
